Used more variables, fixed sort and leaderboard
So the elements in default.vue now change to match what the lobby code and user nickname variables are when the page loads.  Currently those values are hardcoded for testing purposes. The leaderboard now works and has refresh functionality (although this will probably be removed later, mainly just for testing) 

This commit also has tokens which are being removed as the Questions API doesn't do what we previously thought it did. Loading the question in the question box also works!