It works! Answer buttons added, Scoring works
Added the answer buttons, which are now populated with the possible answers. Scoring works too, doesn't update the database yet as I am waiting on the necessary changes to the backend. 

Lifeline buttons are now added but functionality has not been added, as I am waiting for the necessary changes to be made  on the backend.